refactor: criando arquivos para cada classe de coletável
