+++ main.py
@@ -6,72 +6,13 @@ from src.utils.constantes import *
 from src.utils.audio import *
 from src.personagens.professor import Professor
 from src.personagens.aluno import Aluno
+from src.coletaveis.sombrinha import SombrinhaFrevo
+from src.coletaveis.garrafaPitu import GarrafaPitu
+from src.coletaveis.fantasiaCarnaval import FantasiaCarnaval
 
 sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
 
 # --- Classes do Jogo ---
-
-class Item(pygame.sprite.Sprite):
-    """Classe base para todos os itens coletáveis."""
-    def __init__(self, nome, pos, imagens, n_sprites, largura, altura, nova_largura=None, nova_altura=None):
-        super().__init__()
-        self.nome = nome
-        self.image = pygame.Surface((TAMANHO_BLOCO // 2, TAMANHO_BLOCO // 2))
-        self.rect = self.image.get_rect(center=pos)
-        sprite_sheet = pygame.image.load(os.path.join(SPRITES_DIR, imagens)).convert_alpha()
-        self.lista_sprites = []
-        for i in range(n_sprites): # número de sprites
-            sprite = sprite_sheet.subsurface((i * largura, 0), (largura, altura))
-            if nova_altura != None and nova_largura != None:
-                sprite = pygame.transform.scale(sprite, (nova_largura, nova_altura))
-            self.lista_sprites.append(sprite)
-
-        self.index_lista = 0
-
-        # substituindo imagem placeholder
-        self.image = self.lista_sprites[self.index_lista]
-        self.rect = self.image.get_rect(center=self.rect.center)
-
-    def update(self):
-        self.index_lista += 0.15
-        if self.index_lista >= len(self.lista_sprites):
-            self.index_lista = 0
-        self.image = self.lista_sprites[int(self.index_lista)]
-        self.mask = pygame.mask.from_surface(self.image)
-
-    def aplicar_efeito(self, professor):
-        """Método a ser sobrescrito por cada item."""
-        print(f"Item {self.nome} coletado!")
-
-class SombrinhaFrevo(Item):
-    def __init__(self, pos):
-        super().__init__("Sombrinha de Frevo", pos, 'sprite_sheetsombrinha.png', 3, 26, 26)
-        
-    def aplicar_efeito(self, professor):
-        super().aplicar_efeito(professor)
-        som_sombrinha.play()
-        professor.velocidade = professor.velocidade_base * 2
-        professor.tempo_boost = pygame.time.get_ticks() + 5000 # 5 segundos de boost
-
-class GarrafaPitu(Item):
-    def __init__(self, pos):
-        super().__init__("Garrafa de Pitú", pos, 'pitu.png', 2, 60, 190, 10, 32)
-
-    def aplicar_efeito(self, professor):
-        super().aplicar_efeito(professor)
-        professor.drunk = True
-        professor.tempo_drunk = pygame.time.get_ticks() + 7000 # 7 segundos de tontura
-
-class FantasiaCarnaval(Item):
-    def __init__(self, pos):
-        super().__init__("Máscara de Carnaval", pos, 'FantasiaVampiro 19x25.png', 1, 19, 25)
-
-    def aplicar_efeito(self, professor):
-        super().aplicar_efeito(professor)
-        professor.tipo = 'vampiro ' #troca a fantasia
-        professor.invisivel = True
-        professor.tempo_invisivel = pygame.time.get_ticks() + 8000 # 8 segundos de invisibilidade
-        professor.image.set_alpha(128) # Fica semitransparente
 
 class Parede(pygame.sprite.Sprite):
     """Classe para as paredes do labirinto"""
